+++ src/seekwriter.cpp
@@ -59,7 +59,7 @@ bool CSeekWriter::ReadSparseMatrix(const char *fileName,
 	vector<float> rbp_score;
 	rbp_score.resize(maxRank);
 	for(i=0; i<maxRank; i++)
-		rbp_score[i] = (1.0 - rbp_p) * pow(rbp_p, i);
+		rbp_score[i] = (1.0 - rbp_p) * pow(rbp_p, (float)i);
 
 	for(i=0; i<numGenes; i++){
 		utype id, id2;
